app/lambda-functions/get_item/requirements.txt
psycopg2-binary==2.9.7
boto3==1.34.0